live/requirements.txt
streamlit>=1.32
pandas>=2.0
plotly>=5.18
boto3>=1.34
pyyaml>=6.0
